Modularize server resolver files
