+++ packages/server/src/server/rest-api.ts
@@ -6,7 +6,7 @@ import axios from "axios";
 import express from "express";
 import { logger } from "../tools/server-utils";
 import { AxiosUtil, getHostFromNetworkName } from "./axios-utils";
-import { PaginationParams } from "./resolvers";
+import { PaginationParams } from "./resolvers/resolvers";
 import CELO from "./sources/celo";
 import COSMOS_EXTRACTOR from "./sources/cosmos-extractor";
 import OASIS from "./sources/oasis";
